# Scoped roles endpoint query performance