**Address n+1 queries in GET `/roles/?scope=principal`** (47f32e4)
Similar to #172, when roles are scoped to principal, we end up running into
n+1 issues due to needing the policy/role records for the groups for a principal.

This resolves that by added the necessary prefetch lookups on the `get_role_queryset`
calls, and sets up a base query to be reused/extended upon.

We can also setup a `base_query` queryset to build upon in `get_role_queryset`.

---

**Optimize role serializer queries** (b7f299c)
We can remove n+1 queries for roles by prefetching `access` records in the queries
where the serializer is being used on roles.

We don't want to add this to the default query in `group#roles()` because we also use that method for getting the count, in which case we'd be prematurely fetching unnecessary data.

--- 

**Results:** _(same dataset as #172)_
--
**Before:**
```
/roles/?scope=principal
2.79s average response time

/groups/<uuid>/roles/
2.72s average response time
```

**After:**
```
/roles/?scope=principal
729ms average response time (73.87% improvement)

/groups/<uuid>/roles/
978ms average response time (64.04% improvement)
```